Django Rest Frame Work/pov_characters/requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.5
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
whitenoise==5.3.0
